Revert "update model, update ingredient"
This reverts commit 6f3b14d964d240eee56065a95fe0028bc4bdcfcf.

+++ EHM/EHM_API/Repositories/InvoiceRepository.cs
@@ -66,7 +66,7 @@ namespace EHM_API.Repositories
 				Note = order.Discount?.Note,
 				TotalMoney = order.Discount?.TotalMoney,
 				QuantityLimit = order.Discount?.QuantityLimit,
-				Deposits = (decimal)order.Deposits,
+				Deposits = order.Deposits,
 				ItemInvoice = (order.OrderDetails ?? Enumerable.Empty<OrderDetail>()).Select(od => new ItemInvoiceDTO
                 {
                     DishId = od.DishId ?? 0,
